Move ABI related classes to new /abi package.
