+++ rskj-core/src/main/java/co/rsk/peg/BridgeStorageProvider.java
@@ -23,6 +23,7 @@ import co.rsk.bitcoinj.script.Script;
 import co.rsk.config.BridgeConstants;
 import co.rsk.core.RskAddress;
 import co.rsk.crypto.Keccak256;
+import co.rsk.peg.abi.ABICallElection;
 import co.rsk.peg.bitcoin.CoinbaseInformation;
 import co.rsk.peg.federation.Federation;
 import co.rsk.peg.federation.PendingFederation;
