+++ rskj-core/src/main/java/co/rsk/peg/abi/ABICallSpec.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package co.rsk.peg;
+package co.rsk.peg.abi;
 
 import com.google.common.primitives.SignedBytes;
