+++ rskj-core/src/main/java/co/rsk/peg/abi/ABICallVoteResult.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package co.rsk.peg;
+package co.rsk.peg.abi;
 
 /**
  * Immutable representation of the result of a vote
