+++ rskj-core/src/test/java/co/rsk/peg/performance/FederationChangeTest.java
@@ -22,6 +22,8 @@ import co.rsk.bitcoinj.core.BtcECKey;
 import co.rsk.bitcoinj.store.BtcBlockStore;
 import co.rsk.core.RskAddress;
 import co.rsk.peg.*;
+import co.rsk.peg.abi.ABICallElection;
+import co.rsk.peg.abi.ABICallSpec;
 import co.rsk.peg.federation.FederationMember;
 import co.rsk.peg.federation.PendingFederation;
 import org.ethereum.core.Repository;
